remove useless admin client interface
