+++ containers/jersey2-routing/src/main/java/com/yahoo/gondola/container/AdminClient.java
@@ -7,39 +7,38 @@ import com.yahoo.gondola.container.client.ShardManagerClient;
 import java.util.List;
 import java.util.Map;
 
-public class AdminClient implements AdminInterface {
+public class AdminClient {
 
     private String serviceName;
     private Config config;
     private ShardManagerClient shardManagerClient;
-
 
     public AdminClient(String serviceName, ShardManagerClient shardManagerClient) {
         this.serviceName = serviceName;
         this.shardManagerClient = shardManagerClient;
     }
 
-    @Override
+
     public void setServiceName(String serviceName) throws AdminException {
         this.serviceName = serviceName;
     }
 
-    @Override
+
     public String getServiceName() throws AdminException {
         return serviceName;
     }
 
-    @Override
+
     public Config getConfig() throws AdminException {
         return config;
     }
 
-    @Override
+
     public void setConfig(Config config) throws AdminException {
         this.config = config;
     }
 
-    @Override
+
     public void splitShard(String fromShardId, String toShardId) throws AdminException {
         Range<Integer> range = lookupSplitRange(fromShardId, toShardId);
         assignBuckets(fromShardId, toShardId, range);
@@ -47,7 +46,7 @@ public class AdminClient implements AdminInterface {
 
     private void assignBuckets(String fromShardId, String toShardId, Range<Integer> range) {
         shardManagerClient.allowObserver(fromShardId, toShardId);
-        for (int i = 0; i < 3 ; i++) {
+        for (int i = 0; i < 3; i++) {
             try {
                 shardManagerClient.startObserving(toShardId, fromShardId);
                 // assignBucket is a atomic operation executing on leader at fromShard,
@@ -64,7 +63,7 @@ public class AdminClient implements AdminInterface {
         return Range.closed(1, 5);
     }
 
-    @Override
+
     public void mergeShard(String fromShardId, String toShardId) throws AdminException {
         Range<Integer> range = lookupMergeRange(fromShardId, toShardId);
         assignBuckets(fromShardId, toShardId, range);
@@ -72,32 +71,67 @@ public class AdminClient implements AdminInterface {
 
     private Range<Integer> lookupMergeRange(String fromShardId, String toShardId) {
         // TODO: implement
-        return Range.closed(1,5);
+        return Range.closed(1, 5);
     }
 
-    @Override
+
     public void enable(Target target, String targetId) throws AdminException {
 
     }
 
-    @Override
+
     public void disable(Target target, String targetId) throws AdminException {
 
     }
 
-    @Override
+
     public Map<Target, List<Stat>> getStats(Target target)
         throws AdminException {
         return null;
     }
 
-    @Override
+
     public void enableTracing(Target target, String targetId) throws AdminException {
 
     }
 
-    @Override
+
     public void disableTracing(Target target, String targetId) throws AdminException {
 
+    }
+
+    enum Target {
+        HOST, SHARD, SITE, STORAGE, ALL
+    }
+
+    class Stat {
+
+    }
+
+    class HostStat extends Stat {
+
+    }
+
+    class StorageStat extends Stat {
+
+    }
+
+    class ShardStat extends Stat {
+
+    }
+
+    class AdminException extends Exception {
+
+        ErrorCode errorCode;
+    }
+
+    enum ErrorCode {
+        CONFIG_NOT_FOUND(10000);
+
+        private int code;
+
+        ErrorCode(int code) {
+            this.code = code;
+        }
     }
 }
